requirements.txt
fastapi[all]==0.95.0
SQLAlchemy[asyncio]==2.0.9
alembic==1.10.3
python-dotenv==1.0.0
aiohttp==3.8.4
asyncpg==0.27.0
pydantic~=1.10.7
python-jose[cryptography]
passlib[bcrypt]
python-multipart